requirements.txt
# =========================================================================================
# LearnAI 深度学习训练框架 - 依赖清单
# 版本: 3.0
# Python要求: >=3.10
#
# 注意: Web服务部署使用Java（Jakarta EE + Jetty + ONNX），不依赖Python Flask
# =========================================================================================

# =========================
# 深度学习框架
# =========================
tensorflow==2.16.1          # 深度学习主框架 (支持CPU/GPU，按需选择)
                            # GPU版本: pip install tensorflow[and-cuda]
                            # CPU版本: pip install tensorflow-cpu

# =========================
# 科学计算 & 数据处理
# =========================
numpy>=1.26.0               # 数值计算基础库
pandas>=2.2.0               # 表格数据处理（CSV加载）
scipy>=1.11.0               # 科学计算工具（时序增强、插值）

# =========================
# 配置 & 日志
# =========================
pyyaml>=6.0.1               # YAML配置文件解析
colorama>=0.4.6             # 彩色终端日志输出
tqdm>=4.66.1                # 训练/数据加载进度条

# =========================
# 网络通信（强化学习）
# =========================
requests>=2.31.0            # HTTP客户端（与环境服务器通信）

# =========================
# 可视化
# =========================
matplotlib>=3.9.0           # 绘制训练曲线（loss/accuracy）
seaborn>=0.13.0             # 高级数据可视化（可选）

# =========================
# 机器学习工具（可选）
# =========================
scikit-learn>=1.3.0         # 聚类评估、数据预处理
                            # 用于: 轮廓系数、Davies-Bouldin指数等

# =========================
# 模型导出（可选）
# =========================
tf2onnx>=1.15.0             # TensorFlow → ONNX 模型导出
                            # 用于Java服务部署

# =========================
# 开发工具
# =========================
jupyter>=1.0.0              # Jupyter Notebook支持
ipykernel>=6.25.0           # Jupyter内核
pytest>=8.0.0               # 单元测试框架
pytest-cov>=4.1.0           # 测试覆盖率

# =========================
# 代码质量（可选）
# =========================
# black>=23.0.0             # 代码格式化
# flake8>=6.0.0             # 代码检查
# mypy>=1.5.0               # 类型检查
# pylint>=2.17.0            # 代码分析

# =========================================================================================
# 安装说明
# =========================================================================================
#
# 【快速安装 - 最小依赖】（推荐入门）
#   pip install tensorflow numpy pandas pyyaml colorama requests
#
# 【完整安装 - 所有功能】（推荐生产）
#   pip install -r requirements.txt
#
# 【GPU支持】
#   pip install tensorflow[and-cuda]==2.16.1
#   # 需要CUDA 12.x和cuDNN 8.9
#
# 【CPU优化版】
#   pip install tensorflow-cpu==2.16.1
#
# 【国内镜像加速】
#   pip install -i https://pypi.tuna.tsinghua.edu.cn/simple -r requirements.txt
#
# =========================================================================================
# 功能对应的依赖
# =========================================================================================
#
# ✅ 核心功能（监督/强化/自监督学习）
#    → tensorflow, numpy, pandas, pyyaml, colorama, requests
#
# 🎨 数据可视化
#    → matplotlib, seaborn
#
# 📊 聚类评估
#    → scikit-learn
#
# 📦 ONNX模型导出（用于Java部署）
#    → tf2onnx
#
# 🔬 开发和测试
#    → jupyter, pytest
#
# ☕ Java服务部署
#    → 不需要额外Python依赖（使用Jakarta EE + Jetty + ONNX Runtime）
#    → 只需导出ONNX模型: pip install tf2onnx
#
# =========================================================================================
# 版本兼容性
# =========================================================================================
#
# TensorFlow 2.16.1:
#   - Python: 3.9-3.12
#   - NumPy: 1.23.0-1.26.x
#   - CUDA: 12.3 (GPU版本)
#   - cuDNN: 8.9
#
# 测试平台:
#   - Ubuntu 22.04 LTS ✅
#   - macOS 13+ (Apple Silicon/Intel) ✅
#   - Windows 10/11 ✅
#
# =========================================================================================
# 可选的高级功能依赖（按需安装）
# =========================================================================================
#
# 【高级数据增强】
#   pip install albumentations>=1.3.0
#   # 更强大的图像增强库（支持更多变换）
#
# 【分布式训练】
#   pip install horovod>=0.28.0
#   # 多GPU分布式训练支持
#
# 【模型压缩】
#   pip install tensorflow-model-optimization>=0.7.0
#   # 量化、剪枝、蒸馏
#
# 【实验跟踪】
#   pip install tensorboard>=2.16.0
#   # TensorBoard可视化
#   pip install mlflow>=2.10.0
#   # MLflow实验管理
#
# 【云存储】
#   pip install boto3>=1.34.0
#   # AWS S3存储
#   pip install google-cloud-storage>=2.14.0
#   # Google Cloud Storage
#
# =========================================================================================
